+++ handshake-nodejs-reporters/packages/handshake-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # handshake-utils
 
+## 0.4.2
+
+### Patch Changes
+
+- fixed
+
 ## 0.4.1
 
 ### Patch Changes